fix(train): 修复训练中的梯度爆炸问题并添加诊断工具
- 调整配置参数：评估间隔从1000改为800，批次大小从16减至12
- 增加Mel损失权重从25到45，特征匹配损失权重从0.5增至1.0
- 减少保留检查点数量从24到8，禁用音量增强功能
- 在音量增强中限制音量变化范围避免数值爆炸
- 添加梯度裁剪参数设置为100.0
- 新增analyze_gradients.py用于分析TensorBoard梯度范数
- 新增diagnose_training_data.py用于检查训练数据问题
- 更新训练文件列表包含更多数据样本

+++ vdecoder/nsf_hifigan/models.py
@@ -148,7 +148,9 @@ class SineGen(torch.nn.Module):
         rand_ini[:, 0] = 0
         rad_values[:, 0, :] = rad_values[:, 0, :] + rand_ini
         is_half = rad_values.dtype is not torch.float32
-        tmp_over_one = torch.cumsum(rad_values.double(), 1)  # % 1  #####%1意味着后面的cumsum无法再优化
+        # Intel XPU 优化：使用 float32 而非 double，避免 FP64 fallback
+        # cumsum 在 float32 下精度足够（rad_values 范围 0-1）
+        tmp_over_one = torch.cumsum(rad_values.float(), 1)
         if is_half:
             tmp_over_one = tmp_over_one.half()
         else:
@@ -163,8 +165,9 @@ class SineGen(torch.nn.Module):
         tmp_over_one_idx = (tmp_over_one[:, 1:, :] - tmp_over_one[:, :-1, :]) < 0
         cumsum_shift = torch.zeros_like(rad_values)
         cumsum_shift[:, 1:, :] = tmp_over_one_idx * -1.0
-        rad_values = rad_values.double()
-        cumsum_shift = cumsum_shift.double()
+        # Intel XPU 优化：使用 float32 而非 double
+        rad_values = rad_values.float()
+        cumsum_shift = cumsum_shift.float()
         sine_waves = torch.sin(torch.cumsum(rad_values + cumsum_shift, dim=1) * 2 * np.pi)
         if is_half:
             sine_waves = sine_waves.half()
